requirements.txt
python-telegram-bot==21.6
python-dateutil==2.9.0.post0
pytz==2024.2
